requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose
bcrypt
dnspython
requests
Pillow
